requirements.txt
Pillow
scipy
ipykernel
torch
torchvision
scikit-learn
matplotlib
torch
torchvision
fastapi[standard]
torchaudio --index-url https://download.pytorch.org/whl/cu117
sqlalchemy
asyncpg
boto3
python-dotenv
boto3_extensions